requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
python-dotenv
psycopg2-binary==2.9.9
PyJWT
passlib[bcrypt]
